No issue: rename geo -> geofire.
This is a more appropriate name because it's a port of geofire-js. This will
allow me to write fn that are geo-related but keep it separate from the port.
